% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COMBO_EM_data.R
\docType{data}
\name{COMBO_EM_data}
\alias{COMBO_EM_data}
\title{Example data for COMBO-EM}
\format{
\subsection{\code{who}}{

A list with the elements
\describe{
\item{Y}{True outcome (Y)}
\item{Ystar}{Observed outcome (Y*)}
\item{x_matrix}{Matrix of predictors (X)}
\item{z_matrix}{Matrix of adjustors for misclassification (Z)}
\item{true_beta}{True beta}
\item{true_gamma}{True gamma (misclassification)}
\item{pistar}{Individual probability of misclassification}
}
}
}
\usage{
COMBO_EM_data
}
\description{
Example data for COMBO-EM
}
\keyword{datasets}
