% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expit.R
\name{expit}
\alias{expit}
\title{Expit function}
\usage{
expit(x)
}
\arguments{
\item{x}{A numeric value or vector to compute the expit function on.}
}
\value{
\code{expit} returns the result of the function
\eqn{f(x) = \frac{\exp\{x\}}{1 + \exp\{x\}}} for a given \code{x}.
}
\description{
\eqn{\frac{\exp\{x\}}{1 + \exp\{x\}}}
}
\examples{
\dontrun{
x <- c(1, 2, 3)

expit(x)

exp(x) / (1 + exp(x))
}
}
