% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSAC_data.R
\docType{data}
\name{LSAC_data}
\alias{LSAC_data}
\title{Example data from The Law School Admissions Council's (LSAC) National Bar Passage Study (Linda Wightman, 1998)}
\format{
A dataframe 39 columns, including background and demographic information, as well
as if the candidates passed the bar exam to become lawyers in the USA.
}
\source{
\url{https://www.kaggle.com/datasets/danofer/law-school-admissions-bar-passage/data?select=bar_pass_prediction.csv}
}
\usage{
LSAC_data
}
\description{
Example data from The Law School Admissions Council's (LSAC) National Bar Passage Study (Linda Wightman, 1998)
}
\examples{
\dontrun{
data("LSAC_data")
head(LSAC_data)
}
}
\keyword{datasets}
