% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_pistarjj_compute.R
\name{mean_pistarjj_compute}
\alias{mean_pistarjj_compute}
\title{Compute the Mean Conditional Probability of Correct Classification, by True Outcome Across all Subjects}
\usage{
mean_pistarjj_compute(pistar_matrix, j, sample_size)
}
\arguments{
\item{pistar_matrix}{A numeric matrix of conditional probabilities obtained from
the internal function \code{pistar_compute_for_chains}. Rows of the matrix correspond to
each subject and to each observed outcome category. Columns of the matrix
correspond to each true, latent outcome category.}

\item{j}{An integer value representing the true outcome category to compute
the average conditional probability of correct classification for.
\code{j} can take on values \code{1} and \code{2}.}

\item{sample_size}{An integer value specifying the number of observations in the sample.}
}
\value{
\code{mean_pistarjj_compute} returns a numeric value equal to the average
conditional probability \eqn{P(Y^* = j | Y = j, Z)} across all subjects.
}
\description{
Compute the Mean Conditional Probability of Correct Classification, by True Outcome Across all Subjects
}
