% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_picker.R
\name{model_picker}
\alias{model_picker}
\title{Select a Binary Outcome Misclassification Model for a Given Prior}
\usage{
model_picker(prior)
}
\arguments{
\item{prior}{A character string specifying the prior distribution for the
\eqn{\beta} and \eqn{\gamma} parameters. Options are \code{"t"},
\code{"uniform"}, \code{"normal"}, or \code{"dexp"} (double Exponential, or Weibull).}
}
\value{
\code{model_picker} returns a character string specifying the binary
outcome misclassification model to be turned into a .BUG file and used
for MCMC estimation with \code{rjags}.
}
\description{
Select a Binary Outcome Misclassification Model for a Given Prior
}
