% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_picker_2stage.R
\name{model_picker_2stage}
\alias{model_picker_2stage}
\title{Select a Two-Stage Binary Outcome Misclassification Model for a Given Prior}
\usage{
model_picker_2stage(prior)
}
\arguments{
\item{prior}{A character string specifying the prior distribution for the
\eqn{\beta}, \eqn{\gamma}, and \eqn{\delta} parameters. Options are \code{"t"},
\code{"uniform"}, \code{"normal"}, or \code{"dexp"} (double Exponential, or Weibull).}
}
\value{
\code{model_picker} returns a character string specifying the two-stage binary
outcome misclassification model to be turned into a .BUG file and used
for MCMC estimation with \code{rjags}.
}
\description{
Select a Two-Stage Binary Outcome Misclassification Model for a Given Prior
}
