% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive_model_picker_2stage.R
\name{naive_model_picker_2stage}
\alias{naive_model_picker_2stage}
\title{Select a Naive Two-Stage Regression Model for a Given Prior}
\usage{
naive_model_picker_2stage(prior)
}
\arguments{
\item{prior}{A character string specifying the prior distribution for the naive
\eqn{\beta} parameters. Options are \code{"t"},
\code{"uniform"}, \code{"normal"}, or \code{"dexp"} (double Exponential, or Weibull).}
}
\value{
\code{naive_model_picker_2stage} returns a character string specifying the
logistic regression model to be turned into a .BUG file and used
for MCMC estimation with \code{rjags}.
}
\description{
Select a Naive Two-Stage Regression Model for a Given Prior
}
