% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VPRAI_synthetic_data.R
\docType{data}
\name{VPRAI_synthetic_data}
\alias{VPRAI_synthetic_data}
\title{Synthetic example data of pretrial failure risk factors and outcomes, VPRAI recommendations, and judge decisions}
\format{
A dataframe 1990 columns, including defendant race, risk factors, VPRAI
recommendations, judge decisions, and pretrial failure outcomes.
}
\usage{
VPRAI_synthetic_data
}
\description{
Synthetic example data of pretrial failure risk factors and outcomes, VPRAI recommendations, and judge decisions
}
\examples{
\dontrun{
data("VPRAI_synthetic_data")
head(VPRAI_synthetic_data)
}
}
\keyword{datasets}
