% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COMBO_EM_data.R
\docType{data}
\name{COMBO_EM_data}
\alias{COMBO_EM_data}
\title{Test data for the COMBO_EM function}
\format{
A list containing 6 variables for 1000 observations:
\describe{
\item{Y}{The true outcome variable}
\item{Ystar}{The observed outcome variable}
\item{x_matrix}{A matrix of predictor values in the true outcome mechanism}
\item{z_matrix}{A matrix of predictor values in the observed outcome mechanism}
\item{true_beta}{Beta parameter values used for data generation in the true outcome mechanism}
\item{true_gamma}{Gamma parameter values used for data generation in the observed outcome mechanism}
}
}
\usage{
COMBO_EM_data
}
\description{
A dataset for testing the \code{COMBO_EM} function, generated from the
\code{COMBO_data} function.
}
\examples{
\dontrun{
data("COMBO_EM_data")
head(COMBO_EM_data)
}

}
\keyword{datasets}
