% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_switch_2stage.R
\name{label_switch_2stage}
\alias{label_switch_2stage}
\title{Fix Label Switching in MCMC Results from a Binary Outcome Misclassification Model}
\usage{
label_switch_2stage(chain_matrix, dim_x, dim_z, dim_v, n_cat)
}
\arguments{
\item{chain_matrix}{A numeric matrix containing the posterior samples for all
parameters in a given MCMC chain. \code{chain_matrix} must be a named
object (i.e. each parameter must be named as \code{beta[j, p]}, \code{gamma[k,j,p]},
or \code{delta[l,k,j,p]}).}

\item{dim_x}{An integer specifying the number of columns of the design matrix of the true outcome mechanism, \code{X}.}

\item{dim_z}{An integer specifying the number of columns of the design matrix of the second-stage observation mechanism, \code{V}.}

\item{n_cat}{An integer specifying the number of categorical values that the true outcome, \eqn{Y},
the first-stage observed outcome, \eqn{Y^*}, and the second-stage observed
outcome \eqn{\tilde{Y}} can take.}
}
\value{
\code{label_switch_2stage} returns a named matrix of MCMC posterior samples for
all parameters after performing label switching according the following pattern:
all \eqn{\beta} terms are multiplied by -1, all \eqn{\gamma} and \eqn{\delta} terms are "swapped"
with the opposite \code{j} index.
}
\description{
Fix Label Switching in MCMC Results from a Binary Outcome Misclassification Model
}
\examples{
\dontrun{
set.seed(123)
n <- 1000
x_mu <- 0
x_sigma <- 1
z_shape <- 1

beta <- matrix(c(1, 2), ncol = 1)
gamma <- matrix(c(1, 2, 3, 4), nrow = 2, byrow = FALSE)

my_data <- COMBO_data(sample_size = n,
                      x_mu = x_mu, x_sigma = x_sigma,
                      z_shape = z_shape,
                      beta = beta, gamma = gamma)

obs_Y = my_data[["obs_Y"]]
X = my_data[["x_design_matrix"]]
Z = my_data[["z_design_matrix"]]

example_chain1 <- matrix(c(rnorm(n, mean = beta[1,1]),
                           rnorm(n, mean = beta[2,1]),
                           rnorm(n, mean = gamma[1,1]),
                           rnorm(n, mean = gamma[1,2]),
                           rnorm(n, mean = gamma[2,1]),
                           rnorm(n, mean = gamma[2,2])),
                           nrow = n, byrow = FALSE)
colnames(example_chain1) <- c("beta[1,1]", "beta[1,2]",
                              "gamma[1,1,1]", "gamma[1,2,1]",
                              "gamma[1,1,2]", "gamma[1,2,2]")

label_switch_chain1 <- label_switch(example_chain1, ncol(X), ncol(Z), n_cat = 2)

head(example_chain1)
head(label_switch_chain1)
}
}
