% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perfect_sensitivity_EM.R
\name{perfect_sensitivity_EM}
\alias{perfect_sensitivity_EM}
\title{EM-Algorithm Estimation of the Binary Outcome Misclassification Model while Assuming Perfect Sensitivity}
\usage{
perfect_sensitivity_EM(
  Ystar,
  Z,
  X,
  start,
  beta0_fixed = NULL,
  weights = NULL,
  expected = TRUE,
  tolerance = 1e-07,
  max_em_iterations = 1500
)
}
\arguments{
\item{Ystar}{A numeric vector of indicator variables (1, 0) for the observed
outcome \code{Y*}. The reference category is 0.}

\item{Z}{A numeric matrix of covariates in the true outcome mechanism.
\code{Z} should not contain an intercept.}

\item{X}{A numeric matrix of covariates in the observation mechanism.
\code{X} should not contain an intercept.}

\item{start}{Numeric vector of starting values for parameters in the true
outcome mechanism (\eqn{\theta}) and the observation mechanism (\eqn{\beta}), respectively.}

\item{beta0_fixed}{Optional numeric vector of values of the observation mechanism
intercept to profile over. If a single value is entered, this corresponds to
fixing the intercept at the specified value. The default is \code{NULL}.}

\item{weights}{Optional vector of row-specific weights used for selection bias
adjustment. The default is \code{NULL}.}

\item{expected}{A logical value indicating whether or not to calculate the
covariance matrix via the expected Fisher information matrix. The default is \code{TRUE}.}

\item{tolerance}{A numeric value specifying when to stop estimation, based on
the difference of subsequent log-likelihood estimates. The default is \code{1e-7}.}

\item{max_em_iterations}{An integer specifying the maximum number of
iterations of the EM algorithm. The default is \code{1500}.}
}
\value{
\code{perfect_sensitivity_EM} returns a list containing nine elements.
The elements are detailed in \code{?SAMBA::obsloglikEM} documentation. Code
is adapted from the \code{SAMBA::obsloglikEM} function.
}
\description{
EM-Algorithm Estimation of the Binary Outcome Misclassification Model while Assuming Perfect Sensitivity
}
\examples{
\dontrun{
set.seed(123)
n <- 1000
x_mu <- 0
x_sigma <- 1
z_shape <- 1

true_beta <- matrix(c(1, -.5), ncol = 1)
true_gamma <- matrix(c(1, 10, .5, -3), nrow = 2, byrow = FALSE)

my_data <- COMBO_data(sample_size = n,
                      x_mu = x_mu, x_sigma = x_sigma,
                      z_shape = z_shape,
                      beta = true_beta, gamma = true_gamma)

Ystar = ifelse(my_data[["obs_Y"]] == 1, 1, 0)
x_matrix = matrix(my_data[["x"]], ncol = 1)
z_matrix = matrix(my_data[["z"]], ncol = 1)

starting_values <- rep(1,6)
beta_start <- matrix(starting_values[1:2], ncol = 1)
gamma_start <- matrix(starting_values[3:4], ncol = 1, nrow = 2, byrow = FALSE)

perfect_sensitivity_results <- perfect_sensitivity_EM(Ystar,
                                                      Z = x_matrix, X = z_matrix,
                                                      start = c(beta_start, gamma_start),
                                                      beta0_fixed = NULL)
perfect_sensitivity_results$param
}
}
\references{
Beesley, L. and Mukherjee, B. (2020).
Statistical inference for association studies using electronic health records:
Handling both selection bias and outcome misclassification.
Biometrics, 78, 214-226.
}
